Reduces excessive floating bits:
 - Convert x/y position to integers, because we don't need sub-pixel level positioning
 - Created a new type that converts a float64 to just 2 decimal places when JSON encoded for the rotation (angle)

This makes the data a lot cleaner and drops a LOT of extra digits out of every packet for every object!

+++ src/server/player/player.go
@@ -109,7 +109,7 @@ func (p *NetworkPlayer) ApplyBehavior(deltaTime float64, spawnerPipeline chan sh
 	for _, key := range p.HeldKeys {
 		if key == "B" && p.bombCount > 0 && p.bombTime <= 0 {
 			p.bombCount--
-			newBomb := bomb.NewBomb(p.X, p.Y)
+			newBomb := bomb.NewBomb(float64(p.X), float64(p.Y))
 			select {
 			case spawnerPipeline <- newBomb:
 			default:
